+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Add ``gatk-split-ncigar`` process
 - Add process ``gatk-variant-filtration``
 - Add support for ``Python`` 3.10
+- Add workflow ``workflow-rnaseq-variantcalling``
 
 Changed
 -------